+++ collector-amazon-ivs/src/main/java/com/bitmovin/analytics/amazon/ivs/AmazonIvsPlayerCollector.kt
@@ -1,6 +1,7 @@
 package com.bitmovin.analytics.amazon.ivs
 
 import android.content.Context
+import android.os.Handler
 import com.amazonaws.ivs.player.Player
 import com.bitmovin.analytics.BitmovinAnalytics
 import com.bitmovin.analytics.DefaultCollector
@@ -42,7 +43,8 @@ internal class AmazonIvsPlayerCollector(analyticsConfig: AnalyticsConfig, contex
     ): PlayerAdapter {
         val featureFactory: FeatureFactory = AmazonIvsPlayerFeatureFactory(analytics)
         val playerContext = IvsPlayerContext(player)
-        val stateMachine = PlayerStateMachine.Factory.create(analytics, playerContext)
+        val handler = Handler(analytics.context.mainLooper)
+        val stateMachine = PlayerStateMachine.Factory.create(analytics, playerContext, handler)
 
         val playbackService = PlaybackService(stateMachine)
         val playbackManipulator = PlaybackEventDataManipulator(player)